refactor: Split file for each sections
- Before: Each screen have many sections but code of section is not split with code of screen
- After: Split code of sections to each file for each section and import to screen file
