+++ app/src/main/java/timeswap/application/ui/screens/features/home/sections/UserInfoSection.kt
@@ -1,0 +1,48 @@
+package timeswap.application.ui.screens.features.home.sections
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import coil.compose.rememberAsyncImagePainter
+import timeswap.application.R
+
+@Composable
+fun UserInfoSection(fullName: String, avatarUrl: String) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(bottom = 16.dp),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Column {
+            Text(text = "Hello", fontSize = 20.sp, fontWeight = FontWeight.Light)
+            Text(text = "$fullName.", fontSize = 28.sp, fontWeight = FontWeight.Bold)
+        }
+        Image(
+            painter = if (avatarUrl.isNotEmpty()) rememberAsyncImagePainter(avatarUrl) else painterResource(
+                R.drawable.default_avatar
+            ),
+            contentDescription = "Avatar",
+            modifier = Modifier
+                .size(50.dp)
+                .clip(CircleShape),
+            contentScale = ContentScale.Crop
+        )
+    }
+}
