+++ app/src/main/java/timeswap/application/ui/screens/features/jobs/JobScreen.kt
@@ -60,11 +60,13 @@ import timeswap.application.data.entity.JobList
 import timeswap.application.viewmodel.JobListUiState
 import timeswap.application.viewmodel.JobListViewModel
 import timeswap.application.shared.utils.CommonFunction
+import timeswap.application.ui.screens.features.jobs.sections.FilterSection
+import timeswap.application.ui.screens.features.jobs.sections.HeaderSearchSection
+import timeswap.application.ui.screens.features.jobs.sections.JobListSection
 
 @Composable
 fun JobScreen(navController: NavController, viewModel: JobListViewModel) {
     val uiState by viewModel.uiState.collectAsState()
-    var isSwiping by remember { mutableStateOf(false) }
 
     Scaffold(
         bottomBar = { BottomNavigationBar(navController) }
@@ -81,50 +83,10 @@ fun JobScreen(navController: NavController, viewModel: JobListViewModel) {
                 )
         ) {
             // Header Search Section
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(250.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Image(
-                    painter = painterResource(id = R.drawable.background_job),
-                    contentDescription = null,
-                    modifier = Modifier.fillMaxSize(),
-                    contentScale = ContentScale.Crop
-                )
-                Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                    SearchBar()
-                    Spacer(modifier = Modifier.height(15.dp))
-                    LocationSearchBar()
-                }
-            }
+            HeaderSearchSection()
 
             // Filter Section
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Box(
-                    modifier = Modifier
-                        .size(30.dp)
-                        .clip(CircleShape)
-                        .background(Color(0xFFFFA726)),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Icon(
-                        painter = painterResource(id = R.drawable.ic_filter),
-                        contentDescription = "Filter",
-                        modifier = Modifier.size(24.dp),
-                        tint = Color.White
-                    )
-                }
-                FilterChip("Tất Cả Danh Mục", true)
-                FilterChip("Công Nghệ", false)
-                FilterChip("Dịch Vụ", false)
-            }
+            FilterSection()
 
             // Job List Section with Swipe Gesture
             when (uiState) {
@@ -152,151 +114,12 @@ fun JobScreen(navController: NavController, viewModel: JobListViewModel) {
 
                 is JobListUiState.Success -> {
                     val successState = uiState as JobListUiState.Success
-                    val jobList = successState.jobList
-                    val currentPage = successState.pageIndex
-                    val totalPages = successState.totalPages
-
-                    Column(modifier = Modifier.fillMaxSize()) {
-                        LazyColumn(
-                            contentPadding = PaddingValues(22.dp),
-                            modifier = Modifier.pointerInput(Unit) {
-                                detectHorizontalDragGestures { _, dragAmount ->
-                                    if (dragAmount < -100 && currentPage < totalPages) {
-                                        isSwiping = true
-                                        viewModel.nextPage()
-                                        isSwiping = false
-                                    } else if (dragAmount > 100 && currentPage > 1) {
-                                        isSwiping = true
-                                        viewModel.previousPage()
-                                        isSwiping = false
-                                    }
-                                }
-                            }
-                        ) {
-                            items(jobList) { job ->
-                                JobCard(job)
-                                Spacer(modifier = Modifier.height(25.dp))
-                            }
-                        }
-                        Box(
-                            modifier = Modifier
-                                .fillMaxWidth(),
-                            contentAlignment = Alignment.Center
-                        ) {
-                            Text(
-                                text = "Page $currentPage / $totalPages",
-                                fontSize = 16.sp,
-                                fontWeight = FontWeight.Bold,
-                                color = Color.Gray
-                            )
-                        }
-                    }
+                    JobListSection(successState, viewModel)
                 }
             }
         }
     }
 }
-
-@Composable
-fun SearchBar() {
-    var textState by remember { mutableStateOf(TextFieldValue("")) }
-    Row(
-        modifier = Modifier
-            .fillMaxWidth(0.9f)
-            .clip(RoundedCornerShape(12.dp))
-            .background(Color.White)
-            .padding(12.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Icon(
-            painter = painterResource(id = R.drawable.ic_search),
-            contentDescription = "Search",
-            modifier = Modifier.size(20.dp)
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-        BasicTextField(
-            value = textState,
-            onValueChange = { textState = it },
-            textStyle = LocalTextStyle.current.copy(fontSize = 16.sp),
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-@Composable
-fun LocationSearchBar() {
-    var locationState by remember { mutableStateOf(TextFieldValue("")) }
-    Row(
-        modifier = Modifier
-            .fillMaxWidth(0.9f)
-            .clip(RoundedCornerShape(12.dp))
-            .background(Color.White)
-            .padding(12.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Icon(
-            painter = painterResource(id = R.drawable.ic_location),
-            contentDescription = "Location",
-            modifier = Modifier.size(20.dp),
-            tint = Color(0xFFFFA726)
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-        BasicTextField(
-            value = locationState,
-            onValueChange = { locationState = it },
-            textStyle = LocalTextStyle.current.copy(fontSize = 16.sp),
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-@Composable
-fun FilterChip(text: String, selected: Boolean, color: Color = Color.LightGray) {
-    Box(
-        modifier = Modifier
-            .clip(RoundedCornerShape(12.dp))
-            .background(if (selected) Color(0xFFFFA726) else color)
-            .padding(horizontal = 12.dp, vertical = 8.dp)
-    ) {
-        Text(text, color = Color.White, fontSize = 14.sp)
-    }
-}
-
-@Composable
-fun JobCard(job: JobList) {
-    Card(
-        modifier = Modifier.fillMaxWidth().height(140.dp),
-        shape = RoundedCornerShape(16.dp)
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                AsyncImage(
-                    model = job.ownerAvatarUrl,
-                    contentDescription = null,
-                    modifier = Modifier
-                        .size(40.dp)
-                        .clip(CircleShape),
-                    contentScale = ContentScale.Crop
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Column {
-                    Text(text = job.title, fontSize = 18.sp, fontWeight = FontWeight.Bold)
-                }
-            }
-            Spacer(modifier = Modifier.height(8.dp))
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                FilterChip(job.category.categoryName, false, color = Color(0xFF2196F3))
-                FilterChip(job.ward.name, false, color = Color(0xFF4CAF50))
-            }
-            Spacer(modifier = Modifier.height(8.dp))
-            Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
-                Text(CommonFunction.formatDateFromUTC(job.createdAt), fontSize = 12.sp, color = Color.Gray)
-                Text("${job.fee}VNĐ", fontSize = 16.sp, fontWeight = FontWeight.Bold, color = Color.Black)
-            }
-        }
-    }
-}
-
 
 @Preview
 @Composable
