+++ app/src/main/java/timeswap/application/ui/screens/features/jobs/sections/FilterSection.kt
@@ -1,0 +1,62 @@
+package timeswap.application.ui.screens.features.jobs.sections
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import timeswap.application.R
+
+@Composable
+fun FilterSection() {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(16.dp),
+        horizontalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
+        Box(
+            modifier = Modifier
+                .size(30.dp)
+                .clip(CircleShape)
+                .background(Color(0xFFFFA726)),
+            contentAlignment = Alignment.Center
+        ) {
+            Icon(
+                painter = painterResource(id = R.drawable.ic_filter),
+                contentDescription = "Filter",
+                modifier = Modifier.size(24.dp),
+                tint = Color.White
+            )
+        }
+        FilterChip("Tất Cả Danh Mục", true)
+        FilterChip("Công Nghệ", false)
+        FilterChip("Dịch Vụ", false)
+    }
+}
+
+@Composable
+fun FilterChip(text: String, selected: Boolean, color: Color = Color.LightGray) {
+    Box(
+        modifier = Modifier
+            .clip(RoundedCornerShape(12.dp))
+            .background(if (selected) Color(0xFFFFA726) else color)
+            .padding(horizontal = 12.dp, vertical = 8.dp)
+    ) {
+        Text(text, color = Color.White, fontSize = 14.sp)
+    }
+}
