+++ app/src/main/java/timeswap/application/ui/screens/features/jobs/sections/HeaderSearchSection.kt
@@ -1,0 +1,108 @@
+package timeswap.application.ui.screens.features.jobs.sections
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.material3.Icon
+import androidx.compose.material3.LocalTextStyle
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import timeswap.application.R
+
+@Composable
+fun HeaderSearchSection() {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(250.dp),
+        contentAlignment = Alignment.Center
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.background_job),
+            contentDescription = null,
+            modifier = Modifier.fillMaxSize(),
+            contentScale = ContentScale.Crop
+        )
+        Column(horizontalAlignment = Alignment.CenterHorizontally) {
+            SearchBar()
+            Spacer(modifier = Modifier.height(15.dp))
+            LocationSearchBar()
+        }
+    }
+}
+
+@Composable
+fun SearchBar() {
+    var textState by remember { mutableStateOf(TextFieldValue("")) }
+    Row(
+        modifier = Modifier
+            .fillMaxWidth(0.9f)
+            .clip(RoundedCornerShape(12.dp))
+            .background(Color.White)
+            .padding(12.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Icon(
+            painter = painterResource(id = R.drawable.ic_search),
+            contentDescription = "Search",
+            modifier = Modifier.size(20.dp)
+        )
+        Spacer(modifier = Modifier.width(8.dp))
+        BasicTextField(
+            value = textState,
+            onValueChange = { textState = it },
+            textStyle = LocalTextStyle.current.copy(fontSize = 16.sp),
+            modifier = Modifier.fillMaxWidth()
+        )
+    }
+}
+
+@Composable
+fun LocationSearchBar() {
+    var locationState by remember { mutableStateOf(TextFieldValue("")) }
+    Row(
+        modifier = Modifier
+            .fillMaxWidth(0.9f)
+            .clip(RoundedCornerShape(12.dp))
+            .background(Color.White)
+            .padding(12.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Icon(
+            painter = painterResource(id = R.drawable.ic_location),
+            contentDescription = "Location",
+            modifier = Modifier.size(20.dp),
+            tint = Color(0xFFFFA726)
+        )
+        Spacer(modifier = Modifier.width(8.dp))
+        BasicTextField(
+            value = locationState,
+            onValueChange = { locationState = it },
+            textStyle = LocalTextStyle.current.copy(fontSize = 16.sp),
+            modifier = Modifier.fillMaxWidth()
+        )
+    }
+}
