+++ app/src/main/java/timeswap/application/ui/screens/features/profiles/sections/ProfileHeaderSection.kt
@@ -1,0 +1,124 @@
+package timeswap.application.ui.screens.features.profiles.sections
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.navigation.NavController
+import coil.compose.AsyncImage
+import timeswap.application.R
+import timeswap.application.data.entity.User
+import timeswap.application.ui.screens.core.navigation.SettingsDestination
+
+@Composable
+fun ProfileHeaderSection(userProfile : User?, navController: NavController? = null) {
+    val subscriptionText = when (userProfile?.subscriptionPlan) {
+        1 -> "Standard"
+        2 -> "Premium"
+        else -> "Basic"
+    }
+
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(295.dp)
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.background_job),
+            contentDescription = null,
+            modifier = Modifier.fillMaxSize(),
+            contentScale = ContentScale.Crop
+        )
+
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(horizontal = 25.dp, vertical = 40.dp),
+            horizontalAlignment = Alignment.Start
+        ) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+            ) {
+                AsyncImage(
+                    model = userProfile?.avatarUrl ?: "drawable/default_avatar.png",
+                    contentDescription = "Profile Image",
+                    modifier = Modifier
+                        .size(90.dp)
+                        .background(Color.White, shape = CircleShape)
+                )
+
+                IconButton(
+                    onClick = {  navController?.navigate(SettingsDestination.route) },
+                    modifier = Modifier.size(40.dp).padding(top = 0.dp)
+                ) {
+                    Icon(
+                        painter = painterResource(id = R.drawable.ic_setting),
+                        contentDescription = "Settings",
+                        tint = Color.White
+                    )
+                }
+            }
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+
+            Column(
+                verticalArrangement = Arrangement.spacedBy(6.dp),
+                horizontalAlignment = Alignment.Start
+            ) {
+                Text(
+                    text = userProfile?.fullName ?: "Unknown",
+                    fontSize = 20.sp,
+                    fontWeight = FontWeight.Bold,
+                    color = Color.White
+                )
+                Text(
+                    text = userProfile?.fullLocation ?: "Unknown",
+                    fontSize = 14.sp,
+                    color = Color.White.copy(alpha = 0.8f)
+                )
+            }
+        }
+
+        Row(
+            modifier = Modifier
+                .align(Alignment.BottomEnd)
+                .padding(end = 16.dp, bottom = 99.dp)
+                .background(Color.White.copy(alpha = 0.3f), shape = RoundedCornerShape(16.dp))
+                .padding(horizontal = 12.dp, vertical = 6.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(subscriptionText, color = Color.White, fontSize = 14.sp, fontWeight = FontWeight.Medium)
+            Spacer(modifier = Modifier.width(6.dp))
+            Icon(
+                painter = painterResource(id = R.drawable.ic_profile),
+                contentDescription = "Edit",
+                tint = Color.White
+            )
+        }
+    }
+}
